+++ scripts/legacyMacros.js
@@ -463,6 +463,9 @@ export {channelDivinity} from './macros/2014/classFeatures/paladin/channelDivini
 export {guidedStrike} from './macros/2014/classFeatures/paladin/oathOfConquest/guidedStrike.js';
 export {adjustDensity} from './macros/2014/classFeatures/wizard/graviturgyMagic/adjustDensity.js';
 export {gravityWell} from './macros/2014/classFeatures/wizard/graviturgyMagic/gravityWell.js';
+export {chronalShift} from './macros/2014/classFeatures/wizard/chronurgymagic/chronalShift.js';
+export {temporalAwareness} from './macros/2014/classFeatures/wizard/chronurgymagic/temporalAwareness.js';
+export {momentaryStasis} from './macros/2014/classFeatures/wizard/chronurgymagic/momentaryStasis.js';
 // Race Features
 export {aasimarNecroticShroud, aasimarRadiantConsumption, aasimarRadiantSoul, celestialRevelation} from './macros/2014/raceFeatures/aasimar/celestialRevelation.js';
 export {acidBreathWeapon, coldBreathWeapon, fireBreathWeapon, lightningBreathWeapon, poisonBreathWeapon} from './macros/2014/raceFeatures/dragonborn/chromatic/breathWeapon.js';