+++ scripts/macros/2014/classFeatures/wizard/chronurgymagic/temporalAwareness.js
@@ -1,0 +1,5 @@
+export let temporalAwareness = {
+    name: 'Temporal Awareness',
+    version: '1.5.32',
+    rules: 'legacy'
+};